add README.md, fix provider test, and do some more refactoring
