+++ packages/devtools_app/test/goldens/instance_viewer/show_internal_properties.png
