+++ packages/devtools_app/test/inspector_tests/utils/fake_inspector_tree.dart
