+++ packages/devtools_app/test/network_tests/network_table_test.dart
@@ -16,7 +16,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:vm_service/vm_service.dart';
 
 import '../test_data/network.dart';
-import '../test_utils/network_test_utils.dart';
+import 'utils/network_test_utils.dart';
 
 void main() {
   group('NetworkScreen NetworkRequestsTable', () {
